+++ packages/mediawiki-api-base/RELEASENOTES.md
@@ -6,6 +6,7 @@
 - Typing added throughout
 - PSR4 namespacing. Now in `Addwiki\Mediawiki\Api`
 - Client classes, such as `MediaWikiApi`, moved to `Addwiki\Mediawiki\Api`
+- `ApiUser` deprecated in favour of `UserAndPassword` and `UserAndPasswordWithDomain`
 
 ## Version 2.8 (16 February 2021)